🔧 Auto fragment detection.
Detect wether html input is a fragment or a whole document to
enable/disable the "fragment" option for html2pug.

This html2pug's option forces the inclusion of HTML headers even if not
provided in input and vice-versa.

Now doctye or initial `<html>` tag is detected and html2pug's "fragment"
option is adjusted acordingly.